Revert "Refactor gutcheck backened"
