+++ client/components/AppBar.js
@@ -7,11 +7,6 @@ import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
 import MenuIcon from '@mui/icons-material/Menu';
 import LogoutIcon from '@mui/icons-material/LogoutOutlined';
-import FoodBankIcon from '@mui/icons-material/FoodBank';
-import WavesIcon from '@mui/icons-material/Waves';
-import DirectionsRunIcon from '@mui/icons-material/DirectionsRun';
-import BookIcon from '@mui/icons-material/Book';
-
 import { useOktaAuth } from '@okta/okta-react';
 import DashboardIcon from '@mui/icons-material/DashboardOutlined';
 import { useHistory } from 'react-router-dom';
@@ -29,24 +24,9 @@ export default function ButtonAppBar() {
     history.replace('/');
   };
 
-  const toGutcheck = () => {
-    history.replace('/gutcheck');
-  };
-
-  const toMeditation = () => {
-    history.replace('/meditation');
-  };
-
-  const toExercise = () => {
-    history.replace('/exercise');
-  };
-
-  const toJournal = () => {
-    history.replace('/journal');
-  };
-
   return (
-      <AppBar position="static">
+    <Box sx={{ flexGrow: 1, marginTop: 1 }}>
+      <AppBar position="static" sx={{ borderRadius: '3px' }}>
         <Toolbar variant="dense">
           <IconButton
             size="large"
@@ -58,29 +38,20 @@ export default function ButtonAppBar() {
             <MenuIcon />
           </IconButton>
           <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
-            Healthcheck?
-          </Typography>
-            <Button color="inherit" onClick={toGutcheck}>
+            <Button color="inherit" onClick={toHome}>
               Gutcheck
-              <FoodBankIcon sx={{ml: 1}} />
             </Button>
-            <Button color="inherit" onClick={toMeditation}>
-              Meditation
-              <WavesIcon sx={{ml: 1}} />
-            </Button>
-            <Button color="inherit" onClick={toExercise}>
-              Exercise
-              <DirectionsRunIcon sx={{ml: 1}} />
-            </Button>
-            <Button color="inherit" onClick={toJournal}>
-              Journal
-              <BookIcon sx={{ml: 1}} />
-            </Button>
+          </Typography>
+          <Button color="inherit" onClick={toDashboard}>
+            Dashboard
+            <DashboardIcon sx={{ ml: 1 }} />
+          </Button>
           <Button color="inherit" onClick={logout}>
             Logout
             <LogoutIcon sx={{ ml: 1 }} />
           </Button>
         </Toolbar>
       </AppBar>
+    </Box>
   );
 }
