+++ client/components/BarChart.js
