+++ client/hooks/useAuth.js
