--- client/pages/Exercise.js
@@ -1,6 +1,0 @@
-import * as React from 'react';
-
-const Exercise = () => {
-  return <p>Exercise Page</p>;
-};
-export default Exercise;
