--- client/pages/Journal.js
@@ -1,6 +1,0 @@
-import * as React from 'react';
-
-const Journal = () => {
-  return <p>Journal Page</p>;
-};
-export default Journal;
