--- client/pages/Meditation.js
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-const Meditation = () => {
-  return <p>Meditation Page</p>;
-};
-
-export default Meditation;
